Improve logic of server side for finishing of auction event

+++ mobile-demo-n-lyaerd-mvvm-app/app/data/src/main/kotlin/com/lot/mobiledemodata/datasources/network/settings/SettingsNetworkSource.kt
@@ -8,7 +8,7 @@ import retrofit2.http.Path
 
 interface SettingsNetworkSource {
     @GET("/api/settings/{userId}")
-    suspend fun getSettings(@Path("userId") userId: Long): SettingsNetworkModel
+    suspend fun getSettings(@Path("userId") userId: Int): SettingsNetworkModel
 
 //    @POST("/api/settings/notification")
 //    suspend fun updateNotificationSettings()
